requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
pytz>=2024.1
lxml>=5.0.0
beautifulsoup4>=4.12.0
